Cleaning up [skip ci]
