+++ src/plaid/containers/collections.py
@@ -3,7 +3,7 @@
 import copy
 import logging
 from pathlib import Path
-from typing import Optional, Union
+from typing import Optional
 
 import CGNS.MAP as CGM
 import CGNS.PAT.cgnskeywords as CGK
@@ -15,6 +15,10 @@ from CGNS.PAT.cgnsutils import __CHILDREN__, __NAME__
 from plaid.constants import (
     CGNS_ELEMENT_NAMES,
 )
+from plaid.containers.utils import (
+    _check_names,
+    _read_index,
+)
 from plaid.types import CGNSLink, CGNSNode, CGNSTree, Field, Scalar
 from plaid.utils import cgns_helper as CGH
 
@@ -23,98 +27,6 @@ logging.basicConfig(
     format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
     level=logging.INFO,
 )
-
-
-def _read_index(pyTree: list, dim: list[int]):
-    """Read Index Array or Index Range from CGNS.
-
-    Args:
-        pyTree (list): CGNS node which has a child Index to read
-        dim (list): dimensions of the coordinates
-
-    Returns:
-        indices
-    """
-    a = _read_index_array(pyTree)
-    b = _read_index_range(pyTree, dim)
-    return np.hstack((a, b))
-
-
-def _check_names(names: Union[str, list[str]]):
-    """Check that names do not contain invalid character ``/``.
-
-    Args:
-        names (Union[str, list[str]]): The names to check.
-
-    Raises:
-        ValueError: If any name contains the invalid character ``/``.
-    """
-    if isinstance(names, str):
-        names = [names]
-    for name in names:
-        if (name is not None) and ("/" in name):
-            raise ValueError(
-                f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','__')`"
-            )
-
-
-def _read_index_array(pyTree: list):
-    """Read Index Array from CGNS.
-
-    Args:
-        pyTree (list): CGNS node which has a child of type IndexArray_t to read
-
-    Returns:
-        indices
-    """
-    indexArrayPaths = CGU.getPathsByTypeSet(pyTree, ["IndexArray_t"])
-    res = []
-    for indexArrayPath in indexArrayPaths:
-        data = CGU.getNodeByPath(pyTree, indexArrayPath)
-        if data[1] is None:  # pragma: no cover
-            continue
-        else:
-            res.extend(data[1].ravel())
-    return np.array(res, dtype=int).ravel()
-
-
-def _read_index_range(pyTree: list, dim: list[int]):
-    """Read Index Range from CGNS.
-
-    Args:
-        pyTree (list): CGNS node which has a child of type IndexRange_t to read
-        dim (list[str]): dimensions of the coordinates
-
-    Returns:
-        indices
-    """
-    indexRangePaths = CGU.getPathsByTypeSet(pyTree, ["IndexRange_t"])
-    res = []
-
-    for indexRangePath in indexRangePaths:  # Is it possible there are several ?
-        indexRange = CGU.getValueByPath(pyTree, indexRangePath)
-
-        if indexRange.shape == (3, 2):  # 3D  # pragma: no cover
-            for k in range(indexRange[:, 0][2], indexRange[:, 1][2] + 1):
-                for j in range(indexRange[:, 0][1], indexRange[:, 1][1] + 1):
-                    global_id = (
-                        np.arange(indexRange[:, 0][0], indexRange[:, 1][0] + 1)
-                        + dim[0] * (j - 1)
-                        + dim[0] * dim[1] * (k - 1)
-                    )
-                    res.extend(global_id)
-
-        elif indexRange.shape == (2, 2):  # 2D  # pragma: no cover
-            for j in range(indexRange[:, 0][1], indexRange[:, 1][1]):
-                for i in range(indexRange[:, 0][0], indexRange[:, 1][0]):
-                    global_id = i + dim[0] * (j - 1)
-                    res.append(global_id)
-        else:
-            begin = indexRange[0]
-            end = indexRange[1]
-            res.extend(np.arange(begin, end + 1).ravel())
-
-    return np.array(res, dtype=int).ravel()
 
 
 class SampleScalars:
